added json file with blueprints for the key files
key files will be cf1 and cf2. Hot keys are modified by ctrl. F1 is user name,
F2 is password,and F3 will clear the file contents.

the path for each file will have to be edited to match the current working
idrectory. The json contents will then need to be inserted into
/home/$USER/.config/autokey